Try to complete 05 exercise in flex directory
I didn't even get close to doing this one. The
README.md file said to add divs around a bunch of
elements but I thought it meant individually. I
kind of understand what's going on but there was
no way for me to intuitively figure that out. Not
a great exercise if they don't provide you with
the correct methods or approach to solve it.